File Encryption added

+++ Application_Server/services/processFileService.js
@@ -1,8 +1,10 @@
 const fs = require('fs');
 const splitFile = require('split-file');
+const encryptor = require('file-encryptor');
 
 module.exports = {
     processFiles:(files,callback)=>{
+      let key = 'random';
       let status=true;
       
       let targetDir = "./tmp/"+files.sha1;
@@ -31,45 +33,43 @@ module.exports = {
               status = false;
               console.log("file not copied reason : "+err);
             }
+            else
+            {
+              encryptor.encryptFile(path, files.sha1, key, function(err) {
+                  if(err==null || err==undefined)
+                  {
+                    splitFile.splitFileBySize(path, 512*1024)  // shard copied file
+                        .then((names) => {
 
-            splitFile.splitFileBySize(path, 512*1024)  // shard copied file 
-            .then((names) => {
-              
-              try {
-                
-                fs.unlinkSync(path);
+                          try {
 
-                let h = files.sha1;
-                let s = files.size;
+                            fs.unlinkSync(path);
+                            fs.unlinkSync(files.sha1);
+                            let h = files.sha1;
+                            let s = files.size;
 
-               /* let fileObject = {
-                  h:s
-               } */
-              // let fileObject = {};
-               //fileObject[h]=s;
+                            let x = __dirname+'/../dht.json';
+                            let fileString = JSON.parse(fs.readFileSync(x));
 
-               let x = __dirname+'/../dht.json';
-               let fileString = JSON.parse(fs.readFileSync(x));
-               
-              // fileString.push(fileObject);
-              fileString[h]=s;
-               fs.writeFileSync(x,JSON.stringify(fileString,null,4));
-                 //file removed
-                 callback(targetDir);
+                            // fileString.push(fileObject);
+                            fileString[h]=s;
+                            fs.writeFileSync(x,JSON.stringify(fileString,null,4));
+                            //file removed
+                            callback(targetDir);
 
-              } catch(err) {
-                console.error(err);
-                callback(status);
-              }
-            })
-            .catch((err) => {
-              console.log('File not split reason : ', err);
-              callback(status);
-            });
-
+                          } catch(err) {
+                            console.error(err);
+                            callback(status);
+                          }
+                        })
+                        .catch((err) => {
+                          console.log('File not split reason : ', err);
+                          callback(status);
+                        });
+                  }
+              });
+            }
           });
-
-          
       }
       else
       {
